Changed quantity type to Long

+++ src/main/java/pl/maciejak/charity/entity/Donation.java
@@ -29,7 +29,7 @@ public class Donation {
 
     @Positive
     @NotNull
-    private Integer quantity;
+    private Long quantity;
 
     @NotEmpty
     @ManyToMany(cascade = CascadeType.REMOVE)
